refactor: remove _utils module
