+++ expertsystem/solving/conservation_rules.py
@@ -19,8 +19,8 @@ from typing import Any, Callable, List, Optional, Set, Tuple, Type, Union
 
 import attr
 
-from expertsystem._utils import arange
 from expertsystem.data import Spin
+from expertsystem.solving.combinatorics import arange
 from expertsystem.types import EdgeQuantumNumbers, NodeQuantumNumbers
